+++ .eslintrc.json
@@ -35,12 +35,8 @@
     }
   },
   "ignorePatterns": [
-    "*.js",
-    "!.projenrc.js",
-    "*.d.ts",
     "node_modules/",
-    "*.generated.ts",
-    "coverage"
+    "cypress/"
   ],
   "rules": {
     "prettier/prettier": [
