+++ cypress/integration/site-terms.spec.ts
