+++ cypress/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "lib": ["es2017", "dom"],
+    "types": ["cypress"],
+    "baseUrl": ".",
+    "paths": {
+      "*": ["../src/*"],
+    }
+  },
+  "include": ["**/*.ts", "../**/testIds.ts"]
+}
