+++ src/components/CatalogSearch/index.ts
@@ -1,3 +1,2 @@
 export * from "./CatalogSearch";
-export * from "./constants";
 export * from "./CatalogSearchInputs";
