+++ src/components/SearchModal/testIds.ts
@@ -1,0 +1,5 @@
+import { createTestIds } from "util/createTestIds";
+
+const testIds = createTestIds("searchModal", ["container"] as const);
+
+export default testIds;
