+++ src/views/Home/testIds.ts
@@ -1,9 +1,11 @@
 import { createTestIds } from "util/createTestIds";
 
-export const testIds = createTestIds("home", [
+const testIds = createTestIds("home", [
   "headings",
   "results",
   "nextPageBtn",
   "nextIcon",
   "prevIcon",
 ] as const);
+
+export default testIds;
